Rename pca-new.ipynb to IPCA.ipynb
Incremental PCA is a variant of PCA which allow analyzing the data in a incremental manner, it allow fitting the data in batches which makes it ideal for streaming data.
we used it for anomalies detection by transforming data into lower dimensions before reconstructing it back to its original space and computing the reconstruction error for each data point (which is the Euclidean distance between the original data points and reconstructed one’s). A threshold is set by looping through each percentile of the reconstruction error and finding one that yield better precession score.